# packages/core-contracts/dist/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saudi Casting Marketplace API",
    "version": "1.0.0",
    "description": "API for the Saudi Casting Marketplace, connecting professional hirers with verified talent.\n",
    "contact": {
      "name": "API Support",
      "url": "https://enter.tech/support",
      "email": "support@enter.tech"
    },
    "license": {
      "name": "All Rights Reserved"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://saudi-casting.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "API health and status checks."
    },
    {
      "name": "Auth",
      "description": "User authentication and session management."
    },
    {
      "name": "Media",
      "description": "Media asset management and uploads."
    },
    {
      "name": "Search",
      "description": "Search endpoints."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Check API Health",
        "description": "Performs a health check of the API, including database connectivity. This endpoint is not protected by authentication.\n",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "API is unhealthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "security": []
    },
    "/media/uploads": {
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Request a Media Upload URL",
        "description": "Requests a secure, pre-signed URL to upload a media file directly to the S3-compatible storage.\n",
        "operationId": "requestMediaUpload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pre-signed URL created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error. The request body is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/search/talent": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search for Talent Profiles",
        "description": "Performs a search for talent profiles using a query term and optional filters.\n",
        "operationId": "searchTalent",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number to retrieve.",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "hitsPerPage",
            "in": "query",
            "description": "The number of results per page.",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "filters",
            "in": "query",
            "description": "JSON string of filters to apply.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of talent profiles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TalentSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Creates a new user account.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in a user",
        "description": "Authenticates a user and returns JWT tokens.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful (tokens set as HttpOnly cookies)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "description": "Issues a new access and refresh token.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully (tokens set as HttpOnly cookies)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out a user",
        "description": "Revokes the user's refresh token.",
        "operationId": "logoutUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful (cookies cleared)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "$ref": "#/components/schemas/error"
      },
      "error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "A description of the error that occurred.",
            "example": "Internal Server Error"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
      "RequestUpload": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "The name of the file to be uploaded.",
            "example": "self-tape.mp4"
          },
          "contentType": {
            "type": "string",
            "description": "The MIME type of the file.",
            "example": "video/mp4"
          },
          "size": {
            "type": "integer",
            "description": "The size of the file in bytes.",
            "example": 10485760
          },
          "userId": {
            "type": "string",
            "description": "The CUID of the user uploading the file.",
            "example": "clw5j3j0a0000a4gj8k2h9c7b"
          }
        },
        "required": [
          "filename",
          "contentType",
          "size",
          "userId"
        ]
      },
      "RequestUploadResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri",
            "description": "The pre-signed URL to use for the PUT request to upload the file."
          },
          "assetId": {
            "type": "string",
            "format": "cuid",
            "description": "The ID of the newly created media asset record in the database."
          }
        },
        "required": [
          "ok",
          "uploadUrl",
          "assetId"
        ]
      },
      "TalentSearchResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "hits": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "type": "integer",
            "example": 0
          },
          "nbPages": {
            "type": "integer",
            "example": 1
          },
          "hitsPerPage": {
            "type": "integer",
            "example": 20
          },
          "nbHits": {
            "type": "integer",
            "example": 1
          },
          "query": {
            "type": "string",
            "example": "actor"
          }
        },
        "required": [
          "ok",
          "hits",
          "page",
          "nbPages",
          "hitsPerPage",
          "nbHits",
          "query"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}